extract `api` module from `canon`
